refactor(affix): split files
